+++ metadata.rb
@@ -4,9 +4,10 @@ maintainer_email 'support@hw-ops.com'
 license 'Apache 2.0'
 description 'Replication manager for PostgreSQL'
 
-version '0.1.2'
+version '0.2.2'
 
 depends 'postgresql', '>= 2.0.0'
 depends 'build-essential'
 depends 'database'
 depends 'discovery', '>= 0.2.0'
+depends 'runit'